pyproject.toml
[project]
name = "dj-template"
version = "0.1.0"
description = "Django DRF Base Template Project"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.8",
    "djangorestframework>=3.15.0",
    "djangorestframework-simplejwt>=5.3.0",
    "drf-yasg>=1.21.7",
    "celery>=5.3.0",
    "redis>=5.0.0",
    "gunicorn>=21.2.0",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
    "django-cors-headers>=4.3.0",
    "django-extensions>=4.1",
]
